requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
trafilatura==2.0.0
lxml_html_clean==0.4.3
beautifulsoup4==4.13.4
openai==1.99.9
apscheduler==3.10.4
python-dotenv==1.1.1
